+++ crates/ryft-mlir/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Added support for the `shape` dialect.
 - Added support for the `sparse_tensor` dialect.
 - Added support for the `tensor` dialect.
+- Added support for the `transform` dialect.
 - Added support for the Triton `tt` dialect.
 - Added support for the Mosaic GPU dialect.
 - Added support for the Mosaic TPU dialect.